+++ main/src/message/MessageController.h
@@ -8,6 +8,6 @@ namespace MessageController
     cJSON * get(cJSON *  ob);
     void sendMesageSerial(int key, int value);
     void setup();
-    void triangle_changed_event(uint8_t pressed);
-    void square_changed_event(uint8_t pressed);
+    void triangle_changed_event(int pressed);
+    void square_changed_event(int pressed);
 };
